requirements.txt
fastapi[standard]==0.115.12
pydantic==2.11.5
SQLAlchemy==2.0.41
python-dotenv==1.1.0
Jinja2==3.1.6
redis==6.1.0